We saw that Intersect Compressed wtih function has some inefficiencies. I have fixed that in this diff, added a new benchmark to prove the same. 
Dgraph query also tested. 
LDBC09 Query went from 38 second -> 22 second.
Benchmarks before and after:
```
len: 1, compressed: 28, bytes/int: 28.000000
goos: linux
goarch: amd64
pkg: github.com/dgraph-io/dgraph/algo
cpu: 11th Gen Intel(R) Core(TM) i7-1185G7 @ 3.00GHz
                                                                                          │    in_old    │
                                                                                          │    sec/op    │
ListIntersectCompressBin/compressed:IntersectWith:ratio=0.01:size=100:overlap=0.01:-8       701.9n ± ∞ ¹
ListIntersectCompressBin/compressed:IntersectWith:ratio=0.1:size=100:overlap=0.01:-8        10.27µ ± ∞ ¹
ListIntersectCompressBin/compressed:IntersectWith:ratio=1:size=100:overlap=0.01:-8          62.22µ ± ∞ ¹
ListIntersectCompressBin/compressed:IntersectWith:ratio=10:size=100:overlap=0.01:-8         132.2µ ± ∞ ¹
ListIntersectCompressBin/compressed:IntersectWith:ratio=100:size=100:overlap=0.01:-8        125.2µ ± ∞ ¹
ListIntersectCompressBin/compressed:IntersectWith:ratio=0.01:size=1000:overlap=0.01:-8      535.6n ± ∞ ¹
ListIntersectCompressBin/compressed:IntersectWith:ratio=0.1:size=1000:overlap=0.01:-8       3.268µ ± ∞ ¹
ListIntersectCompressBin/compressed:IntersectWith:ratio=1:size=1000:overlap=0.01:-8         1.227m ± ∞ ¹
ListIntersectCompressBin/compressed:IntersectWith:ratio=10:size=1000:overlap=0.01:-8        1.380m ± ∞ ¹
ListIntersectCompressBin/compressed:IntersectWith:ratio=100:size=1000:overlap=0.01:-8       1.157m ± ∞ ¹
ListIntersectCompressBin/compressed:IntersectWith:ratio=0.01:size=10000:overlap=0.01:-8     4.040µ ± ∞ ¹
ListIntersectCompressBin/compressed:IntersectWith:ratio=0.1:size=10000:overlap=0.01:-8      43.04µ ± ∞ ¹
ListIntersectCompressBin/compressed:IntersectWith:ratio=1:size=10000:overlap=0.01:-8        12.65m ± ∞ ¹
ListIntersectCompressBin/compressed:IntersectWith:ratio=10:size=10000:overlap=0.01:-8       13.99m ± ∞ ¹
ListIntersectCompressBin/compressed:IntersectWith:ratio=100:size=10000:overlap=0.01:-8      12.20m ± ∞ ¹
ListIntersectCompressBin/compressed:IntersectWith:ratio=0.01:size=100000:overlap=0.01:-8    67.35µ ± ∞ ¹
ListIntersectCompressBin/compressed:IntersectWith:ratio=0.1:size=100000:overlap=0.01:-8     612.8µ ± ∞ ¹
ListIntersectCompressBin/compressed:IntersectWith:ratio=1:size=100000:overlap=0.01:-8       129.2m ± ∞ ¹
ListIntersectCompressBin/compressed:IntersectWith:ratio=10:size=100000:overlap=0.01:-8      117.3m ± ∞ ¹
ListIntersectCompressBin/compressed:IntersectWith:ratio=0.01:size=1000000:overlap=0.01:-8   1.061m ± ∞ ¹
ListIntersectCompressBin/compressed:IntersectWith:ratio=0.1:size=1000000:overlap=0.01:-8    6.936m ± ∞ ¹
ListIntersectCompressBin/compressed:IntersectWith:ratio=1:size=1000000:overlap=0.01:-8       1.076 ± ∞ ¹
geomean                                                                                     464.6µ
¹ need >= 6 samples for confidence interval at level 0.95

len: 1, compressed: 29, bytes/int: 29.000000
                                                                                          │  in_updated  │
                                                                                          │    sec/op    │
ListIntersectCompressBin/compressed:IntersectWith:ratio=0.01:size=100:overlap=0.01:-8       380.5n ± ∞ ¹
ListIntersectCompressBin/compressed:IntersectWith:ratio=0.1:size=100:overlap=0.01:-8        1.510µ ± ∞ ¹
ListIntersectCompressBin/compressed:IntersectWith:ratio=1:size=100:overlap=0.01:-8          3.142µ ± ∞ ¹
ListIntersectCompressBin/compressed:IntersectWith:ratio=10:size=100:overlap=0.01:-8         8.595µ ± ∞ ¹
ListIntersectCompressBin/compressed:IntersectWith:ratio=100:size=100:overlap=0.01:-8        42.09µ ± ∞ ¹
ListIntersectCompressBin/compressed:IntersectWith:ratio=0.01:size=1000:overlap=0.01:-8      374.3n ± ∞ ¹
ListIntersectCompressBin/compressed:IntersectWith:ratio=0.1:size=1000:overlap=0.01:-8       1.341µ ± ∞ ¹
ListIntersectCompressBin/compressed:IntersectWith:ratio=1:size=1000:overlap=0.01:-8         27.89µ ± ∞ ¹
ListIntersectCompressBin/compressed:IntersectWith:ratio=10:size=1000:overlap=0.01:-8        56.64µ ± ∞ ¹
ListIntersectCompressBin/compressed:IntersectWith:ratio=100:size=1000:overlap=0.01:-8       424.1µ ± ∞ ¹
ListIntersectCompressBin/compressed:IntersectWith:ratio=0.01:size=10000:overlap=0.01:-8     1.334µ ± ∞ ¹
ListIntersectCompressBin/compressed:IntersectWith:ratio=0.1:size=10000:overlap=0.01:-8      5.688µ ± ∞ ¹
ListIntersectCompressBin/compressed:IntersectWith:ratio=1:size=10000:overlap=0.01:-8        291.6µ ± ∞ ¹
ListIntersectCompressBin/compressed:IntersectWith:ratio=10:size=10000:overlap=0.01:-8       745.2µ ± ∞ ¹
ListIntersectCompressBin/compressed:IntersectWith:ratio=100:size=10000:overlap=0.01:-8      4.405m ± ∞ ¹
ListIntersectCompressBin/compressed:IntersectWith:ratio=0.01:size=100000:overlap=0.01:-8    6.140µ ± ∞ ¹
ListIntersectCompressBin/compressed:IntersectWith:ratio=0.1:size=100000:overlap=0.01:-8     27.98µ ± ∞ ¹
ListIntersectCompressBin/compressed:IntersectWith:ratio=1:size=100000:overlap=0.01:-8       2.962m ± ∞ ¹
ListIntersectCompressBin/compressed:IntersectWith:ratio=10:size=100000:overlap=0.01:-8      7.596m ± ∞ ¹
ListIntersectCompressBin/compressed:IntersectWith:ratio=0.01:size=1000000:overlap=0.01:-8   24.85µ ± ∞ ¹
ListIntersectCompressBin/compressed:IntersectWith:ratio=0.1:size=1000000:overlap=0.01:-8    230.0µ ± ∞ ¹
ListIntersectCompressBin/compressed:IntersectWith:ratio=1:size=1000000:overlap=0.01:-8      30.42m ± ∞ ¹
geomean                                                                                     43.42µ
¹ need >= 6 samples for confidence interval at level 0.95
```
Old:
```
len: 1, compressed: 28, bytes/int: 28.000000
goos: linux
goarch: amd64
pkg: github.com/dgraph-io/dgraph/algo
cpu: 11th Gen Intel(R) Core(TM) i7-1185G7 @ 3.00GHz
BenchmarkListIntersectCompressBin/compressed:IntersectWith:ratio=0.01:size=100:overlap=0.01:-8           1698944               701.9 ns/op

BenchmarkListIntersectCompressBin/compressed:IntersectWith:ratio=0.1:size=100:overlap=0.01:-8             112074             10272 ns/op

BenchmarkListIntersectCompressBin/compressed:IntersectWith:ratio=1:size=100:overlap=0.01:-8                19785             62216 ns/op

BenchmarkListIntersectCompressBin/compressed:IntersectWith:ratio=10:size=100:overlap=0.01:-8                8658            132173 ns/op

BenchmarkListIntersectCompressBin/compressed:IntersectWith:ratio=100:size=100:overlap=0.01:-8               8427            125179 ns/op

BenchmarkListIntersectCompressBin/compressed:IntersectWith:ratio=0.01:size=1000:overlap=0.01:-8          2230539               535.6 ns/op

BenchmarkListIntersectCompressBin/compressed:IntersectWith:ratio=0.1:size=1000:overlap=0.01:-8            352585              3268 ns/op

BenchmarkListIntersectCompressBin/compressed:IntersectWith:ratio=1:size=1000:overlap=0.01:-8                 962           1227124 ns/op

BenchmarkListIntersectCompressBin/compressed:IntersectWith:ratio=10:size=1000:overlap=0.01:-8                879           1380495 ns/op

BenchmarkListIntersectCompressBin/compressed:IntersectWith:ratio=100:size=1000:overlap=0.01:-8               962           1156982 ns/op

BenchmarkListIntersectCompressBin/compressed:IntersectWith:ratio=0.01:size=10000:overlap=0.01:-8          268028              4040 ns/op

BenchmarkListIntersectCompressBin/compressed:IntersectWith:ratio=0.1:size=10000:overlap=0.01:-8            27942             43035 ns/op

BenchmarkListIntersectCompressBin/compressed:IntersectWith:ratio=1:size=10000:overlap=0.01:-8                 86          12651328 ns/op

BenchmarkListIntersectCompressBin/compressed:IntersectWith:ratio=10:size=10000:overlap=0.01:-8                81          13986664 ns/op

BenchmarkListIntersectCompressBin/compressed:IntersectWith:ratio=100:size=10000:overlap=0.01:-8               86          12201208 ns/op

BenchmarkListIntersectCompressBin/compressed:IntersectWith:ratio=0.01:size=100000:overlap=0.01:-8          17497             67347 ns/op

BenchmarkListIntersectCompressBin/compressed:IntersectWith:ratio=0.1:size=100000:overlap=0.01:-8            2047            612767 ns/op

BenchmarkListIntersectCompressBin/compressed:IntersectWith:ratio=1:size=100000:overlap=0.01:-8                 8         129177514 ns/op

BenchmarkListIntersectCompressBin/compressed:IntersectWith:ratio=10:size=100000:overlap=0.01:-8                9         117270228 ns/op

BenchmarkListIntersectCompressBin/compressed:IntersectWith:ratio=0.01:size=1000000:overlap=0.01:-8           975           1060993 ns/op

BenchmarkListIntersectCompressBin/compressed:IntersectWith:ratio=0.1:size=1000000:overlap=0.01:-8            172           6935648 ns/op

BenchmarkListIntersectCompressBin/compressed:IntersectWith:ratio=1:size=1000000:overlap=0.01:-8                1        1075782785 ns/op

```
New:
```
len: 1, compressed: 29, bytes/int: 29.000000
goos: linux
goarch: amd64
pkg: github.com/dgraph-io/dgraph/algo
cpu: 11th Gen Intel(R) Core(TM) i7-1185G7 @ 3.00GHz
BenchmarkListIntersectCompressBin/compressed:IntersectWith:ratio=0.01:size=100:overlap=0.01:-8           3412984               380.5 ns/op

BenchmarkListIntersectCompressBin/compressed:IntersectWith:ratio=0.1:size=100:overlap=0.01:-8             727275              1510 ns/op

BenchmarkListIntersectCompressBin/compressed:IntersectWith:ratio=1:size=100:overlap=0.01:-8               359510              3142 ns/op

BenchmarkListIntersectCompressBin/compressed:IntersectWith:ratio=10:size=100:overlap=0.01:-8              122730              8595 ns/op

BenchmarkListIntersectCompressBin/compressed:IntersectWith:ratio=100:size=100:overlap=0.01:-8              25167             42093 ns/op

BenchmarkListIntersectCompressBin/compressed:IntersectWith:ratio=0.01:size=1000:overlap=0.01:-8          3231374               374.3 ns/op

BenchmarkListIntersectCompressBin/compressed:IntersectWith:ratio=0.1:size=1000:overlap=0.01:-8            866000              1341 ns/op

BenchmarkListIntersectCompressBin/compressed:IntersectWith:ratio=1:size=1000:overlap=0.01:-8               41246             27887 ns/op

BenchmarkListIntersectCompressBin/compressed:IntersectWith:ratio=10:size=1000:overlap=0.01:-8              19537             56635 ns/op

BenchmarkListIntersectCompressBin/compressed:IntersectWith:ratio=100:size=1000:overlap=0.01:-8              2635            424116 ns/op

BenchmarkListIntersectCompressBin/compressed:IntersectWith:ratio=0.01:size=10000:overlap=0.01:-8          795170              1334 ns/op

BenchmarkListIntersectCompressBin/compressed:IntersectWith:ratio=0.1:size=10000:overlap=0.01:-8           197584              5688 ns/op

BenchmarkListIntersectCompressBin/compressed:IntersectWith:ratio=1:size=10000:overlap=0.01:-8               4120            291555 ns/op

BenchmarkListIntersectCompressBin/compressed:IntersectWith:ratio=10:size=10000:overlap=0.01:-8              1358            745215 ns/op

BenchmarkListIntersectCompressBin/compressed:IntersectWith:ratio=100:size=10000:overlap=0.01:-8              261           4404634 ns/op

BenchmarkListIntersectCompressBin/compressed:IntersectWith:ratio=0.01:size=100000:overlap=0.01:-8         164694              6140 ns/op

BenchmarkListIntersectCompressBin/compressed:IntersectWith:ratio=0.1:size=100000:overlap=0.01:-8           42037             27977 ns/op

BenchmarkListIntersectCompressBin/compressed:IntersectWith:ratio=1:size=100000:overlap=0.01:-8               393           2961740 ns/op

BenchmarkListIntersectCompressBin/compressed:IntersectWith:ratio=10:size=100000:overlap=0.01:-8              151           7595967 ns/op

BenchmarkListIntersectCompressBin/compressed:IntersectWith:ratio=0.01:size=1000000:overlap=0.01:-8         46863             24845 ns/op

BenchmarkListIntersectCompressBin/compressed:IntersectWith:ratio=0.1:size=1000000:overlap=0.01:-8           4819            229963 ns/op

BenchmarkListIntersectCompressBin/compressed:IntersectWith:ratio=1:size=1000000:overlap=0.01:-8               40          30424977 ns/op
```